Add spotbugs and quick_build profile
This adds a SpotBugs plugin configuration to the build on an opt-in
basis for modules. It also adds a new profile that can disable all
optional functionality in the interest of producing a build as quickly
as possible. The new profile is activated using the io7m.quick_build
property.

Affects #29
Affects #27