Add reproducibility workarounds for JavaDoc
This adds code to the TrivialJavadoc support class that applies various
workarounds to address reproducibility issues in the platform javadoc
tool. Specifically, it replaces the overview-summary.html file with
the index.html file (because the overview-summary.html file contains
a timestamp, even if -notimestamp is used). It also sorts and repacks
the various search index zip files to ensure that they always produce
the same checksum across builds.